+++ src/ui/atoms/Input.tsx
@@ -58,19 +58,20 @@ function InputComponent({
         type={htmlType}
         hasError={Boolean(fieldError?.message)}
         styles={styles}
-        {...register(name) as any}
+        {...register(name) as const}
       />
     </Container>
   )
 }
 
 const Input = styled.input<any>`
-  border-radius: 10px;
+  border-radius: 7px;
   padding: 0 15px;
   font-family: "Catamaran", sans-serif;
   font-optical-sizing: auto;
   font-weight: 100;
   font-style: normal;
+  font-size: 14px;
 
   background-color: ${({ theme }: any) => theme.colors.NEUTRAL.CARD};
   opacity: ${({ readOnly }: any) => (readOnly ? 0.65 : 1)};